remove api and use cacophonyapi, use enum for trackingstatus
