+++ requirements.txt
@@ -20,3 +20,4 @@ PyGObject~=3.34.0
 psutil~=5.6.7
 portalocker~=1.5.2
 toml~=0.10.0
+cacophonyapi
